Lemonade Change
Given price of one lemon is $5. One person can only buy one lemon but can pay using $5,$10 and $20 bill so we have to return the change as $5 for $10 and $15($5+$10) for $20.
So we will keep track of $5 and $10 bills as we get and check if we can return the correct change to the customer.