Minimum Number of Boats | Greedy | Sorting | TC : O(nlogn)
One Boat can array atmost 2 people and of weight atmost of limit. We have to return minimum number of boats to carry every person. Person[i[<=limit given so we will sort the array, and pick people from start and end of array to get max weight and min weight and check if the sum of weight is less then weight or not if yes then one boat for the two people if not one boat for the person with more weight and increase the pointer for that person to pair it up with the minimum one.